Add helper to export HTML document and add debug tool to unstyled export
