+++ .changeset/spicy-meals-promise.md
@@ -1,0 +1,5 @@
+---
+"@lblod/ember-rdfa-editor": minor
+---
+
+Expose helper method to export document as a HTML page
