+++ addon/components/plugins/html-editor/modal.hbs
@@ -5,6 +5,7 @@
   @closeModal={{this.cancel}}
   @modalOpen={{true}}
   @size="large"
+  @padding="none"
 >
   <:title>{{t "ember-rdfa-editor.html-editor.modal.title"}}</:title>
   <:body>
